Implement dynamic view radius. Close #29. Implement universal chunk range trisect method. Add random color tint.

+++ source/voxelman/packets.d
@@ -70,7 +70,7 @@ struct ClientPositionPacket
 // sent by client after receiving SessionInfoPacket
 struct ViewRadiusPacket
 {
-	size_t viewRadius;
+	int viewRadius;
 }
 
 struct ChunkDataPacket
